requirements.txt
pyautogen
pypdf
openai
python-dotenv
langchain
langgraph
pypdf
openai
python-dotenv
pdfplumber
pandas
pdf2image
fitz
pytesseract